+++ mipengine/controller/algorithm_executor_helpers.py
@@ -8,7 +8,6 @@ from mipengine.node_tasks_DTOs import ColumnInfo
 from mipengine.node_tasks_DTOs import DType
 from mipengine.node_tasks_DTOs import TableData
 from mipengine.node_tasks_DTOs import ImmutableBaseModel
-from mipengine.node_tasks_DTOs import UDFArgumentKind
 
 from mipengine.controller.node_tasks_handler_interface import IQueuedUDFAsyncResult
 
@@ -60,14 +59,6 @@ class TableName:
 
     def __repr__(self):
         return self.full_table_name
-
-
-class Literal(ImmutableBaseModel):
-    value: Any
-    kind = UDFArgumentKind.LITERAL
-
-    class Config:
-        allow_mutation = False
 
 
 class _INode(ABC):
