+++ mipengine/controller/node_tasks_handler_interface.py
@@ -4,13 +4,15 @@ from typing import Any
 from typing import List
 from typing import Tuple
 from typing import Optional
+from typing import Dict
 
 from pydantic import BaseModel
 
+from mipengine.node_tasks_DTOs import ImmutableBaseModel
 from mipengine.node_tasks_DTOs import TableData
 from mipengine.node_tasks_DTOs import TableSchema
-from mipengine.node_tasks_DTOs import UDFPosArguments
-from mipengine.node_tasks_DTOs import UDFKeyArguments
+from mipengine.node_tasks_DTOs import UDFArgument
+from mipengine.node_tasks_DTOs import UDFArgumentKind
 
 
 class IAsyncResult(BaseModel, ABC):
@@ -19,6 +21,14 @@ class IAsyncResult(BaseModel, ABC):
     @abstractmethod
     def get(self, timeout=None):
         pass
+
+
+class UDFPosArguments(ImmutableBaseModel):
+    args: List[UDFArgument]
+
+
+class UDFKeyArguments(ImmutableBaseModel):
+    kwargs: Dict[str, UDFArgument]
 
 
 class IQueuedUDFAsyncResult(IAsyncResult, ABC):
