+++ mipengine/node_tasks_DTOs.py
@@ -92,14 +92,6 @@ class UDFArgument(ImmutableBaseModel):
     value: Any
 
 
-class UDFPosArguments(ImmutableBaseModel):
-    args: List[UDFArgument]
-
-
-class UDFKeyArguments(ImmutableBaseModel):
-    kwargs: Dict[str, UDFArgument]
-
-
 # ~~~~~~~~~~~~~~~~~~~ Exceptions ~~~~~~~~~~~~~~~~~~~~~~ #
